+++ src/CallFrame/FunctionCall.php
@@ -53,21 +53,25 @@ final class FunctionCall implements UserLand
         return $this->functionName;
     }
 
+    #[\Override]
     public function file(): Url
     {
         return $this->file;
     }
 
+    #[\Override]
     public function line(): Line
     {
         return $this->line;
     }
 
+    #[\Override]
     public function arguments(): Sequence
     {
         return $this->arguments;
     }
 
+    #[\Override]
     public function toString(): string
     {
         return "{$this->functionName->toString()}()";