+++ src/Link/ToFile.php
@@ -14,6 +14,7 @@ use Innmind\Url\Url;
  */
 final class ToFile implements Link
 {
+    #[\Override]
     public function __invoke(Url $file, Line $line): Url
     {
         return $file;