+++ tests/StackTraceTest.php
@@ -8,7 +8,7 @@ use Innmind\StackTrace\{
     Throwable,
 };
 use Innmind\Immutable\Sequence;
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class StackTraceTest extends TestCase
 {