+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 4.2.0 - 2025-06-07
+
+### Changed
+
+- Requires `innmind/immutable:~5.16`
+- Requires `innmind/graphviz:~4.0`
+
+### Fixed
+
+- PHP `8.4` deprecations
+
 ## 4.1.0 - 2023-09-17
 
 ### Added